Extract Laik_SliceArray out of Laik_Partitioning
No functional change (hopefully). All tests pass.
Compiles with make and cmake.

This is part of an effort to decouple some logic/structs
around partitionings for seperate usage.
The goal is to be able to run partitioners independent from
partitionings, feeding to temporary slice arrays.

In contrast to a partitioning, a slice array just is
an array of slices assigned to task ids.
It does not know about how it was generated, does not know
about slice filtering etc.