configure/Makefile: make sure tests work with OpenMPI
Requesting more MPI ranks than there are physical cores available
may fail with an error message ("not enough slots") with OpenMPI.
To allow tests to succeed on small machines (e.g. only 2 cores),
we need to allow for oversubscription.

This change to configure tries to detect OpenMPI and sets MPIEXEC
environment variable such that "--oversubscribe" option is passed
(the content of MPIEXEC is used for launching MPI in tests if set).

(we could do something similar for CMake)