Partitionings can store multiple SliceArray's now
By default, when using laik_new_partitioning(),
all slices from the partitioner run will be recorded,
and only this result will be stored as SliceArray.
So nothing changes by default.

However, for the pNodes partitioning in the
propagation2d example, we explicitly request to record
two SliceArray's, both using a slice filter, resulting
typically in less slices:
 (1) record only task-local slices
 (2) record slices intersecting with task-local slices
Array (2) is actually enough for calculating all
transitions done in propagation2d. (1) must be recorded
first because this is required to compute (2).

The same can be used in the LULESH port to reduce memory
consumption.