Fix code for switch statistics to work again
This adds a pass over for action sequences to calculate what
is done in one execution of the sequence (number of sends, copies etc).
Whenever a sequence is executed, LAIK-data related statistics are
updated.

This is light-weight enough to always be on.
Statistics is collected per task, and printed at end with LAIK_LOG=2.