KV Store with MPI backend implementation + tests
Simple flat key/value array, with explicit sync being a
global synchronisation in current world of an LAIK instance.
Multiple KV Stores can be synced independently.

Each process collects new and updated entries in a queue,
which is broadcasted to others in next sync. There is a
check in sync to panic with an "update inconsistency" if
the same entry gets set to different values in different
processes. The user is responsible to ensure this does not
happen.

There are tests for single and MPI backend (1 + 4 processes).
TODO: sync implementation for TCP backend. The All-Gather
used in MPI is not implemented in mini-MPI in the TCP backend.

Old start of a more complex hierarchical key-value store not
removed (yet).