TCP backend: adapt backend_mpi.c as dispatcher for TCPs mini-MPI
The TCP backend always was meant to be very similar to the LAIK
MPI backend, just with its own minimal implementation of MPI using TCP.
Thus, backends/tcp/backend.c started as old snapshot of backend_mpi.c
(even before action sequences got used in the backend API).

This commits just "updates" the TCP backend with the current version
of the MPI backend, e.g. taking advantage of action sequence optimization
passes. However, as "mini-MPI" does not implement ISend/IRecv, the
corresponding actions and pass is removed. For more details, run
  git diff --no-index src/backend-mpi.c src/backends/tcp/backend-tcp.c

We lose one feature of the TCP backend: it had a manual reduce
implementation which took advantage of asynchronous functionality
in the mini-MPI implementation of the TCP backend (using callbacks
instead of MPI request objects as isend/irecv do).
However, the TCP backend is meant not for performance but as
playground for fault tolerance features, as long as official MPI
does not have such features. So this seems not to be a big loss.