Allow callbacks to be called on KVS changes
Use this to update location array for direct access.
Now, querying for a location string does not need to
search in KVS. This is much faster as KVS access needs
convertion of location ID into a string and doing string
search on KVS entries.

This also adds basic functionality to remove KVS entries.
Not tested yet.