Introduce an explicit 'current world' of a LAIK instance
An instance covers a set of processes attached to locations.
When processes (at new locations) are added, new locations are added
to the instance. When processes are removed from an instance, we
still remember their location (the processes may be re-added to the
instance).

The active processes within a LAIK instance are defined by the
current world group. The world can be explicitly shrinked by an
application. Or it gets expanded when new processes join (no
API yet).

At the moment, the world group is used in KVS syncs.
All current examples query the world group directly after laik_init,
which makes sure that all initially started processes are part of
that group.

TODO: Make sure that processes removed from an instance do not
appear in any active group of that instance. This requires correct
reference counting of LAIK objects, which is not done yet.