Allow LAIK spaces to be synced via a KVS for spaces
The goal is to sync LAIK objects automatically on expansion
with newly added processes.

Currently spaces explicitly must be added to the store for sync.
This is in internal API, see tests/src/spacestest.c as example.
At some point, we may enable this automatically for all LAIK objects.
TODO: KVS for other LAIK objects.