API change: laik_map* => laik_get_map*
The laik_map functions before did lazy allocation
of mappings for partitionings, and the user could pass
a layout to influence the allocation.

However, specifying a layout is much too late at this
point, as allocation is done at latest when switching
to a new partitioning (or before, with a reservation).
Also, allocations should not be done at arbitrary points
in an explicit API as LAIK wants to be.

Thus, instead of triggering lazy allocation, an allocation
must exist, and the layout is removed from the signature.
Also, rename laik_map* to laik_get_map* to make it clear
that this is an passive API now, just returning parameters
of already existing mappings.

For consistency, the _1d/2d/3d versions all take a mapping
ID now.

All tests pass, which shows that laik_map was already passive
in all example use cases.