Store allocator per mapping
To allow applications to provide pre-allocated
memory space per mapping without the mapping taking
ownership (and not trying to free the memory when
the mapping is destroyed), Laik must be able to remember
how the memory was allocated for a mapping. For this,
an allocator object is stored per mapping. If the allocator
is set to NULL, Laik does not have ownership and does not
free the memory itself.

Usually, memory for mappings are allocated using the
allocator of the container the mapping is used for.
Containers by default use an allocator calling malloc/free.