Introduce laik_set_map_memory to inject own data
Up to now, data managed by a Laik container always started
from an empty partitioning. To initialize it, one had to switch
to a partitioning which automatically allocated required memory
and query the address of allocated memory from the mappings
(if not using pre-allocated memory from a reservation).

laik_set_map_memory() now allows to inject own memory for
an initial partitioning of a Laik container. The initial partitioning
has to be set before with laik_set_initial_partitioning().
The user must provide 1d memory areas covering the memory required
by the initial partitioning. It is assumed that the layout is dense
lexicographical (other layouts may be supported later).
Laik does not take ownership of the provided memory. The user is
responsible to free the memory.

The final summation in example/vsum.c was modified to use the new
API. Note that we can directly inject memory from the local stack
which already holds the partial sums.
However, after switching to the Master-only partitioning, Laik
may allocated new memory. So to read the final sum, the memory
address must be queried.

TODO: allow to inject own memory into reservations.
FIXME: provided memory may be reused after switches, which may
not be expected by users (e.g. when injecting stack memory).
This needs allocation properties, such as "Reuse allowed".