Providing own memory must cover future switches
If the allocator of a mapping is set to NULL, this
means that the user provided its own memory, prohibiting
re-allocation. This is now checked.

For this, the initialization of the allocator is fixed:
it is set the the default allocator of the data container
when allocating the mapping descriptors.

vsum is modified to make use of the knowledge that the memory
provided for mysum reduction stays valid after switch and it
can be accessed on master without querying allocation address.