Fix layout function returing offset of an index
Before, laik_offset returned offset 0 for index (0/0/0),
which made it difficult to use, as a correction for
the slice covered by the layout was required.

Now, a layout knows the valid range of indexes it covers (the
lex layout creation functions are modified accordingly).
This allows the offset function to return the offset for a
given index. To get the pointer for a given element, the
offset must be scaled by element size and added to start
pointer of a mapping. Note that when a mapping is reused
for a slice inside of the originally covered slice, nothing
changes. To get the address, the unchanged start pointer of
the mapping has to be added to the scaled offset, as the
layout just knows about the original slice for allocated
mapping.