Add copy function to layout interface
Copying a slice between two mappings can be implemented
much faster if the layout of the mappings is the same and
it is known to the implementation.

Thus, it makes sense to have a copy function in the layout
interface, much is ensured to be the same known layout
in both the source and destination mapping.

This commit also provides a generic version of the copy
algorithm, just using the offset function of potentially
different layouts.
This can be used for layouts which do not provide a
specialized copy function (or 0) for laik_init_layout.