Create layouts via layout factory function
When data containers create mappings, instead of
hard-coded call to create a lex layout, use a factory
function instead, which can be reset to a custom factory.
This could create custom layouts for mappings.

In jac3d, with "-l", set a custom factory function which
simply returns lex layouts. A custom layout is not
really possible as the stride getter used relys on lex layout.
No test, as this currently is trivial.