Enable layouts to cover multiple mappings
Add a "mapno" function to layout interface that returns
the map number for a given index in the layout.
Modify "offset" function to require a map number.

This is done in preparation for changing a layout to
not just cover one mapping, but all mappings of a data
container. This allows to let the layout decide how
many mappings we need.

TODO:
* this extended layout functionality is not used yet;
  every mapping still allocations its own layout which
  covers just one mapping (the one generating the layout)
* implementations of copy/pack/unpack layout functions
  currently still expect that every mapping has its own
  layout with covers just this mapping.