Use one layout object for regular mappings
Not yet done for mappings in reservations.

This still has layout sections hardcoded to map to separate mappings.

TODO:
When reusing mappings accross switches, we currently simply use the old
layout. Instead, we should update the according section of the new
layout. Because of that, freeing layouts is not done yet and produces
mem leaks.