Fix problem with using old layouts
If a new layout should reuse an old mapping, it must
be modified to cover the old mapping. This is done
via a new "reuse" function in the layout interface.
This checks if for an required allocation (=mapping)
of a new layout an old mapping can be used. If so,
the parameters for this mapping in the new layout
is adjusted to cover the old mapping.

Now no old layouts are reused in newer mappings, and
we can correctly free old layouts.

This also fixes a wrong assertion in offset_lex.