Some documentation for API calls for elasticity
This removes old and misleading comments.
A process group is immutable. For shrinking/expanding,
a new group has to be created and afterwards, the
world gets switched over to the new group. The old group
object still exists, but collective LAIK functions cannot
be called on the old group afterwards. For processes to
get reactivated, they have to join a LAIK instance again.

Reactivation/growing world is not implemented yet.