TCP2: fix use of receive buffers
Before a ID is assigned to a connection, with separate receive
buffer, one buffer was used for all commands. This could
result in mixing/interleaving of partial commands e.g. in startup,
esp. if a lot of TCP2 LAIK processes are started at the same time.

Solution is to use separate receive buffers directly after
connection. We attach the receive buffer to open fd's. However,
the buffer cannot stay in use after closing the connection, as
a new connection via the same fd may want to use its own receive
buffer. Unfortunately, with the push-back mechanism for data
commands, we may not be able to handle all commands when a
connection is closed. Thus, we have to keep the buffer live by
detaching it from the fd, and moving it to the peer struct.

TODO: get rid of push-back mechanism by allowing sending of data
only if space is available at receiver to consume the data (ie.
the application triggers a receive action for the data). This is
similar to RDMA, where writable memory windows are provided to
a sender.
This can be combined with embedding of binary blobs into the
protocol, allowing direct read()'s into LAIK containers.

This commit makes the TCP2 backend to ignore SIGPIPE signals
by calling signal() for global behavior. Should be made local.