TCP2: fix process IDs with shrinked groups + enable 2 tests
While the MPI backend maintains communicators for shrinked
groups and can use rank numbers as given in LAIK transitions,
with TCP2 backend, we have to map IDs ourself.
The IDs used in TCP2 actually are location IDs, and we have
to map via the locationid array given in the used LAIK group.

Now the SPMV2 shrink test works.