TCP2: now sender must request right to send data
LAIK ensures that there is no need to send spurious data
to a receiver; it always needs to trigger a matching receive.
Exploit this and wait at sender side to get send permission
instead of direct sending.
Thus there is no need for push-back of received data when
matching receive is not yet started; all sent data can be
directly consumed
TODO: get rid of the migration logic for the receive buffer