TCP2: Fix race in startup + test
Registering consists of 2 phases:
(1) master accepts "register"
(2) master asks joining processes to "get ready"

The bug was to accept more "register" wishes than needed for
initial world size, but (2) then was not done. This is fixed
by introducing a 2nd startup state for master.

The test uses the new feature of tcp2run launcher to also
launch "spares", ie. processes not required to reach to
requested world size (-n). This triggered the bug.
As the bug results in hanging processes, we use the "timeout"
command to kill the launched processes if not finished after
expected time (here: a bit more than 1 second, timeout is 2s).