TCP2: fix reduce + add markov tests
The markov2 test from MPI exposed a bug in TCP2:
the reduce process aggregates values in the output
map. But if the reduce process had input data for
the reduction, it did not take it into account.
Fix this by copying from input to output map first.

This also enables the markov* tests, taken from
MPI backend.