TCP2: move sending "id" messages out of "register" handler
In direct response to "register", we only answer with one
"id" message to the registering processes with the new assigned ID.
All other "id" lines (of already-known processes) are moved
to the handshake handlers, allowing for more control and it should
be easier to understand, as the sequence of messages is more obvious.
Currently, we just broadcast the same "id" messages here, so this
is no functional change.
However, in startup, we only start sending "id" lines after enough
registrations are accepted.