TCP2: Provide parent group of world for Joining processes
This allows newly joining processes to calculate partitionings
of old and new world for LAIK data objects, and thus take part
of data exchanges via switching from old to new partitioning.

This required to change the TCP2 startup and resize protocols:
master sends "id" messages to new-comers in 2 batches: before
and after telling new-comers to get ready. In the first batch,
IDs and infos for already existing processes are announced,
and in the 2nd batch the IDs of any new-comers.
From this, new-comers get both old and new world, and the
old world is provided as parent of world after laik_init().