Replace laik_set_map_memory with laik_data_provide_memory
For simply/small data containers, the application may allocate
the memory itself to be used by LAIK, with the benefit that
the application does not have to ask for the address of
dynamically allocated memory by LAIK.

laik_set_map_memory was expected to provide an API for that,
but it only did set the memory to use for a given mapping which has
no allocation yet - this could be reused, but not accross switching
to partitionings without any memory use. This bug become obvious
with example vsum3.
Now, laik_data_provide_memory() is just a simple version of a
reservation - prepareMaps() will use that memory if set.
Note that this only provides memory for map 0.

As laik_data_provide_memory blindly uses the provided memory,
it only should be used if it is known that any local partition
using this memory starts at the same index, with the same layout.