Fix laik_switchto_partitioning to cope with group expansion
For switching to a new partitioning with group change,
we migrated the partitioning with the smaller group first
to the larger group, such that the switch happens in the
context of the same group.

However, this was only implemented for shrinking. Fix this
by creating a union of the old and new group, which may be
one of the original groups with simple shrinking or expansion.
If so, we can simply return the larger group.
However, with TCP2 a resize phase may both add new and remove
some processes in the future. The union implementation can
already cope with this.
As a new union may have 2 parents, LaikGroups now can track
relation to two parents, and this is required for migrating
a slice array.