TCP2: implement removal requests, no tests yet
Process removal requests are triggered in TCP2 backend via
 echo "c L1" | telnet localhost 7777
This asks all processes to be removed which contain "L1" in their
location string. Removal requests are executed in resize()
phases and the resulting process group has the processes removed.

Most important new process states in TCP2 backend:
- ReadyRemove: similar to Ready, able to communicate. However, the
  process got marked for removal in resize(), and is removed from the
  group returned by resize(). The process cannot join another
  resize() phase. It is expected that the application calls
  laik_finalize() and exits. It also can call laik_init() again.
- Dead: processes mark a peer as dead as sign to not communicate with it
  again. Peer processes marked for removal after a resize() can
  communicate regularily (to send their data to Ready processes), but
  they get marked as "Dead" by peers as soon as they enter another
  resize() phase.
Other states are used within the resize() phase.

This commit allows removal of processes with the simple resize example.
Other code is not yet expected to work.
Removal of process LID 0 is not yet working.