jac1d: iteration to decide if reading from data1/2
Before, joining processes started with the same role (read/write)
for data1 and data2. So depending on whether new joining processes
joined in even or uneven iteration, they used the same container
in the new proceses as the existing ones, or not.
Even though it worked (for communication it doesn't matter), this
is confusing. Fix this.