Introduce laik_finish_world_resize
This allows LAIK (and e.g. the TCP2 backend) to free local
resources and mark processes which where started to be removed
as dead.
This is a pure local action. Also, it is optional: another
world resize will finish a previous resize automatically
(as well as a laik_finalize in the removed processes).

Theoretically, this could trigger a sync between master and the
removed processes via the backend, to ensure that the removed
processes actually finished with their resize actions.
However, this does not help a resource manager, as removed
processes still may take some time for cleanup activity. Instead,
a resource manager should wait for processes to terminate and get
a notifcation about that by other means.