Configure: always use python3, no use of distutils
Before, the configure script could be run with Python version 2 and 3.
Now assume that version 3 should be available on all systems with
Python (note that running the configure script is not required).

The distutils package is deprecated in Python 3, showing an
ugly warning with 3.10. Thus replace disutils.spawn.find_executable
with shutil.which from the standard library.