tcp2: always disable Nagle's algorithm
Disabling Nagle is important for fast communication of the
TCP2 backend as it uses push semantics: data is sent to peers
without them replying back, as they enter the next computation
phase as soon as values got updated. As TCP defaults to sending
ACKs delayed (hoping in being able to pickyback them in replys),
and Nagle only sends new data after receiving an ACK, any next
updates would be sent only after delayed ACKS are reveived.

Before, Nagle only was disabled in sockets connecting to the
tcp2 master process, but not in the other direction (ie. at master).

This makes the tcp2 test suite run in 14s instead of 67s locally
on a 2-core Intel CPU on Linux 6.5.