Allow toggling between async and sync
Further work on getting either of these modes to work. The problem is
that there is no ordering between messages, so just because we received
n remote CQ entries, it doesn't mean that the receives were actually the
ones that we were waiting for.

This commit tried using the round number for synchronization. It did not
work, because round number can be different between sender and receiver.