Fix jac1d by adding laik_finish_world_resize
We require malleable code to explicitly tell LAIK when we are
done transfering data between old and new processes to be able
to free the resources and enable process to terminate.
That is, after calling laik_finish_world_resize, the application
cannot access the old group any longer, as it does not longer
exist.